updates Student class: adds helper class methods for working with database
